requirements.txt
streamlit
pdfplumber
pytesseract
pillow
pandas
reportlab
openpyxl
